@@ -9,7 +9,11 @@ const formatName = (nameWithDash) => nameWithDash.replace('-', ' ')
 
 const PokemonPage = ({ previous, next }) => {
   const { name } = useParams()
-  const { data: pokemon, error, isLoading } = useApi(`https://pokeapi.co/api/v2/pokemon/${name}`)
+  const {
+    data: pokemon,
+    error,
+    isLoading
+  } = useApi(`https://pokeapi.co/api/v2/pokemon/${name}`)
 
   if (isLoading) {
     return <LoadingSpinner />
@@ -19,13 +23,18 @@ const PokemonPage = ({ previous, next }) => {
   }
 
   const { type } = pokemon.types.find((type) => type.slot === 1)
-  const stats = pokemon.stats.map((stat) => ({
-    name: formatName(stat.stat.name),
-    value: stat.base_stat
-  })).reverse()
+  const stats = pokemon.stats
+    .map((stat) => ({
+      name: formatName(stat.stat.name),
+      value: stat.base_stat
+    }))
+    .reverse()
   const normalAbility = pokemon.abilities.find((ability) => !ability.is_hidden)
-  const hiddenAbility = pokemon.abilities.find((ability) => ability.is_hidden === true)
+  const hiddenAbility = pokemon.abilities.find(
+    (ability) => ability.is_hidden === true
+  )
 
+  // eslint-disable-next-line no-console
   console.log('hiddenAbility=', hiddenAbility)
   return (
     <>
@@ -35,7 +44,10 @@ const PokemonPage = ({ previous, next }) => {
         {next && <Link to={`/pokemon/${previous.name}`}>Next</Link>}
       </div>
       <div className={`pokemon-page pokemon-type-${type.name}`}>
-        <div className="pokemon-image" style={{ backgroundImage: `url(${pokemon.sprites.front_default})` }} />
+        <div
+          className="pokemon-image"
+          style={{ backgroundImage: `url(${pokemon.sprites.front_default})` }}
+        />
         <div className="pokemon-info">
           <div className="pokemon-name">{pokemon.name}</div>
           <div className="pokemon-stats" data-testid="stats">
@@ -51,8 +63,16 @@ const PokemonPage = ({ previous, next }) => {
             </table>
           </div>
           <div className="pokemon-abilities">
-            {normalAbility && <PokemonAbility abilityName={formatName(normalAbility.ability.name)} />}
-            {hiddenAbility && <PokemonAbility abilityName={formatName(hiddenAbility.ability.name)} />}
+            {normalAbility && (
+              <PokemonAbility
+                abilityName={formatName(normalAbility.ability.name)}
+              />
+            )}
+            {hiddenAbility && (
+              <PokemonAbility
+                abilityName={formatName(hiddenAbility.ability.name)}
+              />
+            )}
           </div>
         </div>
       </div>
